requirements.txt
# Основные
python-telegram-bot==20.7
ccxt==4.2.0
pybit==5.6.2

# База данных
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Анализ данных
pandas==2.1.4
numpy==1.26.2

# Графики (для команды /chart)
matplotlib==3.8.2
mplfinance==0.12.10b0

# Утилиты
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio-throttle==1.0.2

# Мониторинг
psutil==5.9.7

# Логирование
loguru==0.7.2
